+++ test/pdf_combiner_create_pdf_from_multiple_images_test.dart
@@ -1,0 +1,106 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:pdf_combiner/communication/pdf_combiner_platform_interface.dart';
+import 'package:pdf_combiner/pdf_combiner.dart';
+import 'package:pdf_combiner/responses/pdf_combiner_status.dart';
+
+import 'mocks/mock_pdf_combiner_platform.dart';
+
+// Mock platform that simulates an error in the mergeMultiplePDF method.
+class MockPdfCombinerPlatformWithError extends MockPdfCombinerPlatform {
+  @override
+  Future<String?> mergeMultiplePDFs({
+    required List<String> inputPaths,
+    required String outputPath,
+  }) {
+    return Future.error('Simulated Error');
+  }
+}
+
+void main() {
+  group('PdfCombiner  Create PDF From Multiple Images Unit Tests', () {
+    // Test for error handling when file not exist in the createPDFFromMultipleImages method.
+    test('createPDFFromMultipleImages - Error handling (empty inputPaths)',
+        () async {
+      MockPdfCombinerPlatform fakePlatform = MockPdfCombinerPlatform();
+
+      // Replace the platform instance with the mock implementation.
+      PdfCombinerPlatform.instance = fakePlatform;
+
+      // Call the method and check the response.
+      final result = await PdfCombiner.createPDFFromMultipleImages(
+        inputPaths: [],
+        outputPath: 'output/path',
+      );
+
+      // Verify the error result matches the expected values.
+      expect(result.response, null);
+      expect(result.status, PdfCombinerStatus.error);
+      expect(result.message, 'The parameter (inputPaths) cannot be empty');
+    });
+
+    // Test for error handling when file not exist in the createPDFFromMultipleImages method.
+    test('createPDFFromMultipleImages - Error handling (File does not exist)',
+        () async {
+      MockPdfCombinerPlatform fakePlatform = MockPdfCombinerPlatform();
+
+      // Replace the platform instance with the mock implementation.
+      PdfCombinerPlatform.instance = fakePlatform;
+
+      // Call the method and check the response.
+      final result = await PdfCombiner.createPDFFromMultipleImages(
+        inputPaths: ['path1.jpg', 'path2.jpg'],
+        outputPath: 'output/path',
+      );
+
+      // Verify the error result matches the expected values.
+      expect(result.response, null);
+      expect(result.status, PdfCombinerStatus.error);
+      expect(result.message, 'File does not exist: path1.jpg');
+    });
+
+    // Test for error handling when you try to send a file that its not an image in createPDFFromMultipleImages
+    test('createPDFFromMultipleImages - Error handling (File is not an image)',
+        () async {
+      MockPdfCombinerPlatform fakePlatform = MockPdfCombinerPlatform();
+
+      // Replace the platform instance with the mock implementation.
+      PdfCombinerPlatform.instance = fakePlatform;
+
+      // Call the method and check the response.
+      final result = await PdfCombiner.createPDFFromMultipleImages(
+        inputPaths: ['assets/document_1.pdf', 'path2.jpg'],
+        outputPath: 'output/path',
+      );
+
+      // Verify the error result matches the expected values.
+      expect(result.response, null);
+      expect(result.status, PdfCombinerStatus.error);
+      expect(result.message,
+          'Only Image file allowed. File is not an image: assets/document_1.pdf');
+    });
+
+    // Test for success process in createPDFFromMultipleImages
+    test('createPDFFromMultipleImages - success generate PDF', () async {
+      MockPdfCombinerPlatform fakePlatform = MockPdfCombinerPlatform();
+
+      // Replace the platform instance with the mock implementation.
+      PdfCombinerPlatform.instance = fakePlatform;
+
+      // Call the method and check the response.
+      final result = await PdfCombiner.createPDFFromMultipleImages(
+        inputPaths: [
+          'example/assets/image_1.jpeg',
+          'example/assets/image_2.png'
+        ],
+        outputPath: 'output/path',
+      );
+
+      // Verify the error result matches the expected values.
+      expect(result.response, 'Created PDF from Images');
+      expect(result.status, PdfCombinerStatus.success);
+      expect(result.message, 'Processed successfully');
+      expect(result.toString(),
+          'PdfFromMultipleImageResponse{response: ${result.response}, message: ${result.message}, status: ${result.status} }');
+    });
+  });
+}
